+++ index.css
@@ -1,49 +1,3 @@
-.res-logo{
-    height: 90px;
-}
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 5px 25px;
-    border: 1px solid black;
-    height: 100%;
-    background-color: rgb(158, 163, 187);
-}
-
-.header > div > ul > li {
-    list-style-type: none;
-    margin: 10px;
-    font-size: 17px;
-}
-
-.header > div > ul {
-    display: flex;
-}
-
-.card{
-    background-color: #f0f0f0;
-    width: 200px;
-    margin: 10px;
-    padding: 3px;
-}
-
-.card:hover{
-    border: 1px solid black;
-    border-radius: 6px;
-    cursor: pointer;
-    transition: 1.5ms;
-}
-
-.images{
-    width: 100%;
-    height: 140px;
-    object-fit: cover;
-}
-
-.resContainer{
-    display: flex;
-    flex-wrap: wrap;
-    padding: 20px 60px;
-}
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
