+++ index.html
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Swiggy App</title>
-    <link rel="stylesheet" href="index.css">
-</head>
-<body>
-   <div id="root">
+    <link rel="stylesheet" href="index.css" />
+  </head>
+  <body>
+    <div id="root">
       <h1>Not Rendered</h1>
-   </div>
-</body>
-<script type="module" src="App.js"></script>
+    </div>
+  </body>
+  <script type="module" src="/src/App.js"></script>
 </html>
