+++ package.json
@@ -18,11 +18,15 @@
   },
   "homepage": "https://github.com/Naiduchandrasekhar/Namaste-React01#readme",
   "devDependencies": {
+    "@types/react": "^18.2.60",
     "parcel": "^2.11.0",
-    "process": "^0.11.10"
+    "postcss": "^8.4.35",
+    "process": "^0.11.10",
+    "tailwindcss": "^3.4.1"
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0"
   }
 }
