+++ src/components/About.js
@@ -1,0 +1,12 @@
+import ProfileClass from "./ProfileClass"
+
+const About = () => {
+    return(
+        <div>
+            <h1>This is About us page</h1>
+            <ProfileClass name="Chandra sekhar" role="Software Engineer" location="Vizag" />
+        </div>
+    )
+}
+
+export default About
