+++ src/components/Grocery.js
@@ -1,0 +1,9 @@
+const Grocery = () => {
+ return (
+    <div>
+        <h1>This is grocery list</h1>
+        <button>Order now</button>
+    </div>
+ )
+}
+export default Grocery
