+++ src/utils/ThemeContext.js
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+const ThemeContext = createContext({
+    mode: null
+})
+
+export default ThemeContext
