+++ src/utils/constants.js
@@ -1,0 +1,15 @@
+export const CDN_URL =
+  "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
+
+export const LOGO_URL =
+  "https://image.similarpng.com/very-thumbnail/2021/09/Good-food-logo-design-on-transparent-background-PNG.png";
+
+export const ALL_RESTAURANTS_URL = "https://foodfire.onrender.com/api/restaurants?lat=21.1702401&lng=72.83106070000001&page_type=DESKTOP_WEB_LISTING"
+
+export const RESTAURANT_MENU_URL =  "https://foodfire.onrender.com/api/menu?page-type=REGULAR_MENU&complete-menu=true&lat=21.1702401&lng=72.83106070000001&&submitAction=ENTER&restaurantId="
+
+export const MENU_ITEM_IMG = "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/"
+
+export const  EACH_MENU_ITEM = "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_208,h_208,c_fit/"
+
+export const OFFLINE_URL = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRBKWM14zOL7cEZoass_eu20GGIC_JEpdUHXw&usqp=CAU"
