+++ tailwind.config.js
@@ -1,0 +1,9 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: ["./src/**/*.{html,js}",],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
+
